add command line options

+++ Program.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Linq;
+using System.Runtime.InteropServices;
 using dnlib.DotNet;
 using dnlib.DotNet.Emit;
 
@@ -12,10 +13,18 @@ namespace AssemblyCSharpPreprocessor
         {
             if (args.Length < 1)
             {
-                Console.WriteLine("Usage: AssemblyCSharpPreprocessor <input file> [<output file>]");
+                Console.WriteLine("Usage: AssemblyCSharpPreprocessor <input file> [<output file>] [options]");
+                Console.WriteLine("Options: ");
+                Console.WriteLine("\t-d, --delete-unused-classes: delete unused classes (iOSPlugins, SMS)");
+                Console.WriteLine("\t-r, --rename-classes: rename classes (Math => Math2)");
+                Console.WriteLine("\t-c, --change-all-private-members-to-internal: change all access modifier of private members (types, methods, fields, properties, events) to internal");
+                Console.WriteLine("\t-o, --overwrite-file: overwrite input file");
+                Environment.Exit(1);
                 return;
             }
             string inputFile = args[0];
+            CommandLineOptions options = CommandLineOptions.Parse(args);
+            args = args.Where(s => !s.StartsWith("-")).ToArray();
             string outputFile = args.Length > 1 ? args[1] : null;
             byte[] data = File.ReadAllBytes(inputFile);
             ModuleDefMD assemblyCSharp = ModuleDefMD.Load(data);
@@ -24,12 +33,24 @@ namespace AssemblyCSharpPreprocessor
                 Console.WriteLine("Input file is not Assembly-CSharp!");
                 return;
             }
-            DeleteUnusedClasses(assemblyCSharp);
-            RenameClasses(assemblyCSharp);
-            ChangeAllPrivateToInternal(assemblyCSharp);
+            if (options.DeleteUnusedClasses)
+                DeleteUnusedClasses(assemblyCSharp);
+            if (options.RenameClasses)
+                RenameClasses(assemblyCSharp);
+            if (options.ChangeAllPrivateMembersToInternal)
+                ChangeAllPrivateMembersToInternal(assemblyCSharp);
             try
             {
-                assemblyCSharp.Write(outputFile ?? inputFile);
+                if (outputFile != null)
+                {
+                    if (options.OverwriteInputFile)
+                        Console.WriteLine("Not overwriting original file!");
+                    assemblyCSharp.Write(outputFile);
+                }
+                else if (options.OverwriteInputFile)
+                    assemblyCSharp.Write(inputFile);
+                else
+                    assemblyCSharp.Write(Path.Combine(Path.GetDirectoryName(inputFile), Path.GetFileNameWithoutExtension(inputFile) + "-processed" + Path.GetExtension(inputFile)));
             }
             catch (Exception ex)
             {
@@ -37,17 +58,19 @@ namespace AssemblyCSharpPreprocessor
                 if (!File.Exists(inputFile))
                     File.WriteAllBytes(inputFile, data);
             }
-            if (outputFile == null)
+            if (outputFile == null && options.OverwriteInputFile)
                 File.WriteAllBytes(Path.Combine(Path.GetDirectoryName(inputFile), Path.GetFileNameWithoutExtension(inputFile) + "-original" + Path.GetExtension(inputFile)), data);
         }
 
         static void RenameClasses(ModuleDefMD assemblyCSharp)
         {
+            Console.WriteLine("Renaming classes...");
             assemblyCSharp.Find("Math", false).Name = "Math2";
         }
 
         static void DeleteUnusedClasses(ModuleDefMD assemblyCSharp)
         {
+            Console.WriteLine("Deleting unused classes...");
             MethodDef fixedUpdate = assemblyCSharp.Find("Main", false).FindMethod("FixedUpdate");
             for (int i = fixedUpdate.Body.Instructions.Count - 1; i >= 0; i--)
             {
@@ -65,8 +88,9 @@ namespace AssemblyCSharpPreprocessor
                 assemblyCSharp.Types.Remove(iOSPlugins);
         }
 
-        static void ChangeAllPrivateToInternal(ModuleDefMD assemblyCSharp)
+        static void ChangeAllPrivateMembersToInternal(ModuleDefMD assemblyCSharp)
         {
+            Console.WriteLine("Changing access modifiers...");
             foreach (TypeDef type in assemblyCSharp.GetTypes().Where(t => t.GetClosestCompilerGeneratedAttribute() == null))
             {
                 foreach (MethodDef method in type.Methods)
